Receiving callback and related function moved to a new file for class NotificationServerConnection
